requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
async-timeout==5.0.1
attrs==25.3.0
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
propcache==0.3.0
pypika-tortoise==0.5.0
python-dotenv==1.0.1
pytz==2025.1
tortoise-orm==0.24.2
typing_extensions==4.12.2
yarl==1.18.3
